LayoutTests/svg/path-arc-invalid-expected.txt

CONSOLE MESSAGE: Error: Problem parsing d="M0 0 A0 0 0 0 0 0 0 0 0 0"
 If this text is visible the malformed SVG path is handled correctly.
